ship_backend/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==18.9.0
importlib-metadata==6.6.0
Markdown==3.4.3
parameterized==0.9.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
zipp==3.15.0
